+++ src/commands/addFolder.ts
