+++ src/commands/removeFolder.ts
